pyproject.toml
[project]
name = "siem-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.218.1",
    "clickhouse-connect>=0.10.0",
    "requests>=2.32.5",
    "pyyaml>=6.0.0",
    "typing-extensions>=4.15.0",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "vulture>=2.14",
]
